fix: import buildTableHelper[5042]
