+++ policy-service/src/policy-engine/helpers/custom-logic-worker.ts
@@ -2,7 +2,7 @@ import { workerData, parentPort } from 'node:worker_threads';
 import * as mathjs from 'mathjs';
 import * as formulajs from '@formulajs/formulajs'
 
-import { buildTableHelper } from '../helpers/table-field.js';
+import { buildTableHelper } from '../helpers/table-field-core.js';
 
 /**
  * Execute function
